extracted components from plugin-ui
